# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Network monitoring and management platform API.",
        "title": "SubNetree API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/recon/scan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Trigger a new network scan on the given subnet. Returns immediately with scan ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recon"
                ],
                "summary": "Start scan",
                "parameters": [
                    {
                        "description": "Subnet to scan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_recon.ScanRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Scan accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.ScanResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.APIProblem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.APIProblem"
                        }
                    }
                }
            }
        },
        "/recon/scans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of scan results.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recon"
                ],
                "summary": "List scans",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Max results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.ScanResult"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.APIProblem"
                        }
                    }
                }
            }
        },
        "/recon/scans/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single scan result including discovered devices.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recon"
                ],
                "summary": "Get scan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.ScanResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.APIProblem"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.APIProblem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.APIProblem"
                        }
                    }
                }
            }
        },
        "/recon/topology": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the network topology as a graph of nodes and edges.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recon"
                ],
                "summary": "Get topology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_recon.TopologyGraph"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.APIProblem"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_HerbHall_subnetree_pkg_models.APIProblem": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "invalid request body"
                },
                "instance": {
                    "type": "string",
                    "example": "/api/v1/auth/login"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "title": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "type": {
                    "type": "string",
                    "example": "https://subnetree.com/problems/bad-request"
                }
            }
        },
...
        "github_com_HerbHall_subnetree_pkg_models.Device": {
            "type": "object",
            "properties": {
                "agent_id": {
                    "type": "string",
                    "example": "agent-01"
                },
                "custom_fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "device_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.DeviceType"
                        }
                    ],
                    "example": "server"
                },
                "discovery_method": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.DiscoveryMethod"
                        }
                    ],
                    "example": "icmp"
                },
                "first_seen": {
                    "type": "string",
                    "example": "2026-01-10T08:00:00Z"
                },
                "hostname": {
                    "type": "string",
                    "example": "web-server-01"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "ip_addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "last_seen": {
                    "type": "string",
                    "example": "2026-01-15T10:30:00Z"
                },
                "mac_address": {
                    "type": "string",
                    "example": "00:1a:2b:3c:4d:5e"
                },
                "manufacturer": {
                    "type": "string",
                    "example": "Dell Inc."
                },
                "notes": {
                    "type": "string",
                    "example": "Production web server"
                },
                "os": {
                    "type": "string",
                    "example": "Ubuntu 22.04"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.DeviceStatus"
                        }
                    ],
                    "example": "online"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_HerbHall_subnetree_pkg_models.DeviceStatus": {
            "type": "string",
            "enum": [
                "online",
                "offline",
                "degraded",
                "unknown"
            ],
            "x-enum-varnames": [
                "DeviceStatusOnline",
                "DeviceStatusOffline",
                "DeviceStatusDegraded",
                "DeviceStatusUnknown"
            ]
        },
        "github_com_HerbHall_subnetree_pkg_models.DeviceType": {
            "type": "string",
            "enum": [
                "server",
                "desktop",
                "laptop",
                "mobile",
                "router",
                "switch",
                "printer",
                "iot",
                "unknown"
            ],
            "x-enum-varnames": [
                "DeviceTypeServer",
                "DeviceTypeDesktop",
                "DeviceTypeLaptop",
                "DeviceTypeMobile",
                "DeviceTypeRouter",
                "DeviceTypeSwitch",
                "DeviceTypePrinter",
                "DeviceTypeIoT",
                "DeviceTypeUnknown"
            ]
        },
        "github_com_HerbHall_subnetree_pkg_models.DiscoveryMethod": {
            "type": "string",
            "enum": [
                "agent",
                "icmp",
                "arp",
                "snmp",
                "mdns",
                "upnp",
                "mqtt",
                "manual"
            ],
            "x-enum-varnames": [
                "DiscoveryAgent",
                "DiscoveryICMP",
                "DiscoveryARP",
                "DiscoverySNMP",
                "DiscoverymDNS",
                "DiscoveryUPnP",
                "DiscoveryMQTT",
                "DiscoveryManual"
            ]
        },
        "github_com_HerbHall_subnetree_pkg_models.ScanResult": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.Device"
                    }
                },
                "ended_at": {
                    "type": "string",
                    "example": "2026-01-15T10:32:15Z"
                },
                "id": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                },
                "online": {
                    "type": "integer",
                    "example": 8
                },
                "started_at": {
                    "type": "string",
                    "example": "2026-01-15T10:30:00Z"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "subnet": {
                    "type": "string",
                    "example": "192.168.1.0/24"
                },
                "total": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
...
        "internal_recon.ScanRequest": {
            "type": "object",
            "properties": {
                "subnet": {
                    "type": "string",
                    "example": "192.168.1.0/24"
                }
            }
        },
        "internal_recon.TopologyEdge": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "link-001"
                },
                "link_type": {
                    "type": "string",
                    "example": "ethernet"
                },
                "source": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "speed": {
                    "type": "integer",
                    "example": 1000
                },
                "target": {
                    "type": "string",
                    "example": "660f9500-f30c-52e5-b827-557766551111"
                }
            }
        },
        "internal_recon.TopologyGraph": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_recon.TopologyEdge"
                    }
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_recon.TopologyNode"
                    }
                }
            }
        },
        "internal_recon.TopologyNode": {
            "type": "object",
            "properties": {
                "device_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.DeviceType"
                        }
                    ],
                    "example": "server"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "ip_addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "label": {
                    "type": "string",
                    "example": "web-server-01"
                },
                "mac_address": {
                    "type": "string",
                    "example": "00:1a:2b:3c:4d:5e"
                },
                "manufacturer": {
                    "type": "string",
                    "example": "Dell Inc."
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_HerbHall_subnetree_pkg_models.DeviceStatus"
                        }
                    ],
                    "example": "online"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}